requirements.txt
addict==2.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
click==8.1.3
comm==0.1.3
configargparse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flask==2.3.1
fonttools==4.39.3
fsspec==2023.4.0
huggingface-hub==0.14.1
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyterlab-widgets==3.0.7
kdl-py==1.1.0
kiwisolver==1.4.4
markupsafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
open3d==0.17.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
platformdirs==3.4.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
pyrealsense2==2.53.1.4623
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
pyzmq==25.0.2
regex==2023.5.5
scikit-learn==1.2.2
scipy==1.10.1
segment-anything==1.0
sentencepiece==0.1.91
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
timm==0.6.13
tokenizers==0.13.3
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing==3.7.4.3
tzdata==2023.3
wcwidth==0.2.6
werkzeug==2.3.0
widgetsnbextension==4.0.7
zmq==0.0.0
transformations
kinpy>=0.2.2
timm>=0.6.13
wandb
classifier_free_guidance_pytorch
opencv-contrib-python==4.6.0.66
git+https://github.com/openai/CLIP.git
